// src/app/api/model/create-expense-dto.ts
/**
 * Auth API (NestJS + Prisma)
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CreateExpenseDto {
  /**
   * วันที่ (YYYY-MM-DD HH:mm:ss)
   */
  startProjectDate: string;
  type: CreateExpenseDto.TypeEnum;
  amount: number;
  category: string;
  note?: string;
  orderNo: string;
  websiteName: string;
  address?: string;
  telNo?: string;
  customerType: CreateExpenseDto.CustomerEnum;
  withholdingTaxPercent: number;
  withholdingTaxAmount: number;
  serviceFeePercent: number;
  serviceFeeAmount: number;
  expenseItems?: CreateExpenseItemDto[];
}
export namespace CreateExpenseDto {
  export const TypeEnum = {
    Income: 'income',
    Expense: 'expense',
  } as const;
  export type TypeEnum = (typeof TypeEnum)[keyof typeof TypeEnum];
  export const CustomerEnum = {
    Individual: 'individual',
    Company: 'Company',
  } as const;
  export type CustomerEnum = (typeof CustomerEnum)[keyof typeof CustomerEnum];
}

export interface CreateExpenseItemDto {
  name: string;
  amount: number;
}